Split the TodoItem model into separate files
